Due to gofmt the commit is too big
The only change I've made is removing initializeClient function due to it was causing fails.
Even with retries it, the pod cannot be at healthy state and k8s cannot start it.
So, I've removed it and it's ok. There is voltdb_up metric already and added some log.Errorf about what's happening. I think that's how it should be.